backend/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
numpy
python-dotenv
google-generativeai
pydantic